requirements.txt
alembic
flask
flask_cors
openai
psycopg2-binary
python-dotenv
redis
sqlalchemy
waitress